CRUD only for task_user

+++ app/Http/Controllers/TasksController.php
@@ -68,10 +68,17 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        
+        }else {
+            return redirect('/');
+        }
+        
     }
 
     /**
@@ -81,12 +88,19 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
+    {   
         $task = Task::find($id);
         
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        
+        }else {
+            return redirect('/');
+        }
+        
     }
 
     /**
@@ -104,11 +118,20 @@ class TasksController extends Controller
         ]);
         
         $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        
+        if (\Auth::id() === $task->user_id) {
 
-        return redirect('/');
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+            
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        
+        }else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -120,8 +143,17 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            
+            $task->delete();
+            
+            return redirect('/');
+        
+        }else {
+            return redirect('/');
+        }
+        
+        
     }
 }
